Relocate all macros into per-module `internal` objects
Macro implementations now live in a predictable place — an `object
internal` per module (or `protointernal` where a same-package `internal`
already exists in a depended-on module). Migrates the eight remaining
non-conforming macro objects (CssInterpolator/StyleMacros/CssMacros,
HtmlMacros, JsonPointerInterpolator, Jacinta, XPathInterpolator,
YamlPathInterpolator, UrlInterpolator, and the EmailAddress/Hostname
`expand` macros) to match the convention already followed by the rest of
the codebase. No behavioural change.
