--- lib/cataclysm/src/core/cataclysm.StyleMacros.scala
@@ -1,66 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.54.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package cataclysm
-
-import scala.quoted.*
-
-import anticipation.*
-import fulminate.*
-import gossamer.*
-import vacuous.*
-
-// Compile-time machinery behind the `Css.Style(borderWidth = …, color = …)`
-// dynamic constructor. Each named parameter's camelCase label becomes a
-// kebab-case property name; the value's type must have a `CssConvertible`
-// instance (which both renders it and tags it with a value-definition-syntax
-// type); and that type is checked against the property's grammar, so e.g.
-// `Css.Style(color = 4.0*Px)` and `Css.Style(notAProperty = …)` fail to compile.
-object StyleMacros:
-  def style(properties: Expr[Seq[(Label, Any)]])(using Quotes): Expr[Css.Style] =
-    def recur(exprs: Seq[Expr[(Label, Any)]]): List[Expr[(Text, Text)]] = exprs match
-      case '{type key <: Label; ($key: key, $value: value)} +: tail =>
-        val convertible = Expr.summon[value is CssConvertible].getOrElse:
-          halt(m"cataclysm: no CSS value is available for this property's value")
-
-        val name = key.value.getOrElse(halt(m"cataclysm: the property name must be a literal"))
-        val property = name.tt.uncamel.kebab
-        CssMacros.propertyIssue(property, CssMacros.topicOf(convertible)).let(halt(_))
-
-        '{(${Expr(property)}, $convertible.value($value))} :: recur(tail)
-
-      case _ =>
-        Nil
-
-    properties match
-      case Varargs(exprs) => '{Css.Style.of(${Expr.ofList(recur(exprs))})}
-      case _              => '{Css.Style.of(Nil)}
