+++ lib/urticose/src/core/urticose.EmailAddress.scala
@@ -32,15 +32,11 @@
                                                                                                   */
 package urticose
 
-import scala.compiletime.*
-import scala.quoted.*
-
 import anticipation.*
 import contingency.*
 import denominative.*
 import distillate.*
 import fulminate.*
-import gigantism.*
 import gossamer.*
 import hypotenuse.*
 import prepositional.*
@@ -57,19 +53,6 @@ object EmailAddress:
 
   given encodable: EmailAddress is Encodable in Text = _.text
   given showable: EmailAddress is Showable = _.text
-
-  def expand(context: Expr[StringContext]): Macro[EmailAddress] = abortive:
-    val text: Text = context.valueOrAbort.parts.head.tt
-    val address = EmailAddress.parse(text)
-
-    val localPart: Expr[LocalPart] = address.localPart match
-      case LocalPart.Quoted(text)   => '{LocalPart.Quoted(${Expr(text)})}
-      case LocalPart.Unquoted(text) => '{LocalPart.Unquoted(${Expr(text)})}
-
-    address.domain.asMatchable.absolve match
-      case ipv6: Ipv6         => '{EmailAddress(Unset, $localPart, ${Expr(ipv6)})}
-      case hostname: Hostname => '{EmailAddress(Unset, $localPart, ${Expr(hostname)})}
-      case ipv4: Int          => '{EmailAddress(Unset, $localPart, Ipv4(${Expr(ipv4)}))}
 
   def parse(text: Text): EmailAddress raises EmailAddressError =
     val buffer: StringBuilder = StringBuilder()
