+++ lib/urticose/src/core/urticose.Hostname.scala
@@ -38,7 +38,6 @@ import anticipation.*
 import contingency.*
 import denominative.*
 import distillate.*
-import gigantism.*
 import gossamer.*
 import hypotenuse.*
 import prepositional.*
@@ -54,10 +53,6 @@ object Hostname:
   given decodable: Tactic[HostnameError] => Hostname is Decodable in Text = parse(_)
   given encodable: Hostname is Encodable in Text = showable.text(_)
 
-
-  def expand(context: Expr[StringContext]): Macro[Hostname] = abortive:
-    Expr(Hostname.parse(context.valueOrAbort.parts.head.tt))
-
   given toExpr: ToExpr[Hostname]:
     def apply(hostname: Hostname)(using Quotes): Expr[Hostname] =
       val labels = Varargs:
@@ -66,7 +61,7 @@ object Hostname:
 
       '{Hostname($labels*)}
 
-  private def parse(text: Text): Hostname raises HostnameError =
+  private[urticose] def parse(text: Text): Hostname raises HostnameError =
     val builder: TextBuilder = TextBuilder()
 
     def recur(index: Ordinal, dnsLabels: List[DnsLabel]): Hostname = text.at(index) match
